@@ -1,16 +1,16 @@
 #:schema https://json.schemastore.org/pyproject.json
 
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["flit_core >=3.11,<4"]
+build-backend = "flit_core.buildapi"
 
 [project]
 name = "stowage"
 dynamic = ["version"]
 description = "Stow, but in Python, and in a single file"
 license = { file = "LICENSE" }
-authors = [{ name = "Keith Gaughan", email = "keith+stowage@gaughan.ie" }]
-requires-python = ">=3.8"
+authors = [{ name = "Keith Gaughan", email = "k+stowage@stereochro.me" }]
+requires-python = ">=3.11"
 readme = "README.rst"
 classifiers = [
   "Environment :: Console",
@@ -20,9 +20,6 @@ classifiers = [
   "Intended Audience :: End Users/Desktop",
   "Operating System :: OS Independent",
   "Programming Language :: Python",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: Implementation :: CPython",
@@ -37,50 +34,57 @@ Source = "https://github.com/kgaughan/stowage"
 [project.scripts]
 stowage = "stowage:main"
 
-[tool.rye]
-dev-dependencies = ["mypy>=1.10.1"]
-universal = true
-generate-hashes = true
-managed = true
-
-[tool.hatch.build.targets.sdist]
-include = ["requirements.lock", "stowage.py"]
-
-[tool.hatch.version]
-path = "stowage.py"
+[tool.flit.module]
+name = "stowage"
 
 [tool.black]
 line-length = 120
 
 [tool.ruff]
-target-version = "py37"
+target-version = "py311"
 line-length = 120
 src = ["src"]
 
 [tool.ruff.lint]
 select = [
   "A",
+  "ANN",
   "ARG",
   "B",
   "C",
   "DTZ",
   "E",
   "EM",
+  "EXE",
   "F",
   "FBT",
+  "FLY",
+  "FURB",
+  "G",
   "I",
   "ICN",
   "ISC002",
+  "LOG",
   "N",
+  "PERF",
+  "PIE",
   "PLC",
   "PLE",
   "PLR",
   "PLW",
+  "PT",
+  "PYI",
   "Q",
+  "RET",
+  "RSE",
   "RUF",
   "S",
+  "SIM",
+  "SLOT",
   "T",
+  "TC",
   "TID",
+  "TRY",
   "UP",
   "W",
   "YTT",
